% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenameClusters.R
\name{ClusterRename}
\alias{ClusterRename}
\title{New Name Clusters
Create a vector of cluster names. The function takes a csv file with old names in 1st columnn 
and new names in 2nd column. The levels of the output are ordered the same as the CSV file.}
\usage{
ClusterRename(Clusters, ClusterCSV, ClusterOrder = NULL)
}
\arguments{
\item{Clusters}{vector of cluster names (same length as number of cells)}

\item{ClusterCSV}{a CSV file containing old names in 1st columnn and new names in 2nd column}

\item{ClusterOrder}{A vector specifying cluster order for plotting. 
This should be given with the new cluster names. 
Default is NULL. This will plot the cluster in the same order 
as CSV file.}
}
\description{
New Name Clusters
Create a vector of cluster names. The function takes a csv file with old names in 1st columnn 
and new names in 2nd column. The levels of the output are ordered the same as the CSV file.
}
\examples{
ClustersRename(seuratObj@active.ident, "plotCols.csv", ClusterOrder = c("Cluster_3", "Cluster_4", "Cluster_5", "Cluster_1", "Cluster_2" ))

}
