% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DowngradeSeurat.R
\name{DowngradeSeuratObject}
\alias{DowngradeSeuratObject}
\title{Function to downgrade Seurat object from verson 3 to version 2}
\usage{
DowngradeSeuratObject(object, assayUse = "RNA")
}
\arguments{
\item{SeuratObj}{seurat version 3 object}
}
\description{
Function to downgrade Seurat object from verson 3 to version 2
}
\examples{
sObj_V2 <- DowngradeSeuratObject(sObj)
}
