% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSum.R
\name{geneSetSumSCE}
\alias{geneSetSumSCE}
\title{Gene Sum SCE}
\usage{
geneSetSumSCE(SCE, Features, ListName = "ListName",
  Exprs_Type = "scale")
}
\arguments{
\item{SCE}{SCE object}

\item{Features}{Gene list}

\item{ListName}{name for list to be added to coldata (default is ListName)}

\item{Exprs_Type}{either counts or scale.data (default is scale.data)}

\item{Clusters}{Vector indicating cluster identity equal to number of cells in Seurat Object}
}
\description{
Calculate total expression for a list of genes and add them onto a Seurat object
}
\examples{
geneSetSCE(SCE, c("Nrsn1",   "Vsnl1",   "Pcp4",    "Nrgn" ))
}
