% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\docType{class}
\name{jackstraw.data-class}
\alias{jackstraw.data-class}
\alias{jackstraw.data}
\title{The jackstraw.data class}
\description{
The JackStrawData is used to store the results of a JackStraw computation.
}
\section{Slots}{

\describe{
\item{\code{empirical.p.value}}{Empirical p-values}

\item{\code{fake.pc.scores}}{Fake reduction scores}

\item{\code{empirical.p.value.full}}{Empirical p-values on full}
}}

