% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myDimPlot.R
\name{MyDimPlot}
\alias{MyDimPlot}
\title{My Dim Plot function for dimension reduction plots calculated by Seurat. 
More options for plotting different cluster names}
\usage{
MyDimPlot(SeuratObj, Clusters, Reduction.Type = c("umap", "tsne", "pca"),
  Label.Size = 3, Point.Size = 3, Point.Colours = NULL,
  Show.Legend = FALSE)
}
\arguments{
\item{SeuratObj}{seurat object}

\item{Clusters}{Vector indicating cluster identity equal to number of cells in Seurat Object}

\item{Reduction.Type}{Which dimensionality reduction to use (umap, tsne, pca)}

\item{Label.Size}{size for cluster label}

\item{Point.Colours}{vector of colours with length of number of clusters.}

\item{Show.Legend}{Show plot legend. Default is False}
}
\description{
My Dim Plot function for dimension reduction plots calculated by Seurat. 
More options for plotting different cluster names
}
\examples{
clusters = seuratObj@active.ident
myDimPlot(seuratObj, clusters, Reduction.Type = "tsne")
}
