% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots.R
\name{PercentGene}
\alias{PercentGene}
\title{Calculate Percentage of Reads attributed to a certain set of genes and adds them to Seurat Object Metadata.
Eg Percentage of mitochondrial genes}
\usage{
PercentGene(SeuratObj, Pat = "^MT-", GeneList = NULL,
  GeneListName = "mito", Assay = "RNA")
}
\arguments{
\item{Pat}{= Select genes using prefix. Eg select mitochondrial genes using patten "^MT-" (default)}

\item{GeneList}{Select genes using names, row number or a T/F vector (length equal to number of genes in Seurat object)}

\item{GeneListName}{Name of Gene list. Default is mito. Results are added to Seurat object using this name (eg percent.GeneListName)}

\item{Assay}{which Seurat assay to use. Default is "RNA"}
}
\description{
Calculate Percentage of Reads attributed to a certain set of genes and adds them to Seurat Object Metadata.
Eg Percentage of mitochondrial genes
}
\examples{
seuratObj = percentGene(seuratObj, Pat = "^MT")

}
