% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots.R
\name{QCplot2D}
\alias{QCplot2D}
\title{Dot plot showing data from metadata column of Seurat object}
\usage{
QCplot2D(SeuratObj, X, Y, ColourBy = NULL, Xmin = NULL, Xmax = NULL,
  Ymin = NULL, Ymax = NULL)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{X}{Column from Metadata to plot}

\item{Y}{Column from Metadata to plot}

\item{ColourBy}{Column fro Metadata}

\item{Xmin}{plot abline for minimum X}

\item{Xmax}{plot abline for maximum X}

\item{Ymin}{plot abline for minimum Y}

\item{Ymax}{plot abline for maximum Y}
}
\description{
Dot plot showing data from metadata column of Seurat object
}
\examples{
X = "nFeature_RNA"
Y = "nCount_RNA"

PlotQC(seuratObj, X, Y)
}
