% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_plots.R
\name{QCplotHist}
\alias{QCplotHist}
\title{Create cell histograms showing number of genes per cell}
\usage{
QCplotHist(SeuratObj, MetaCol, Xmin = NULL, Xmax = NULL,
  NumberOfBins = NULL)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{MetaCol}{Column from Metadata to plot}

\item{Xmin}{plot abline for minimum X}

\item{Xmax}{plot abline for maximum X}

\item{NumberOfBins}{Number of Bins}
}
\description{
Create cell histograms showing number of genes per cell
}
\examples{
clusters = seuratObj@active.ident
CellHist(seuratObj, MetaCol = "nCount_RNA", NumberOfBins = 10)

}
