% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RenameClusters.R
\name{ReadPlotCols}
\alias{ReadPlotCols}
\title{Read Plot Colours}
\usage{
ReadPlotCols(ClusterCSV, ColNumber = 3, ClusterCol = 2)
}
\arguments{
\item{ClusterCSV}{a CSV file containing cluster colours}

\item{ColNumber}{Column where colours are stored (default is 3)}

\item{ClusterCol}{Column containing cluster names (default is 2)}
}
\description{
Create a vector of cluster names. The function takes a csv file with old names in 1st columnn 
and new names in 2nd column.
The the levels of the output are ordered the same as the CSV file.
}
\examples{
ReadPlotCols("plotcolours.csv")

}
