% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{SCEtoSeurat}
\alias{SCEtoSeurat}
\title{Convert SCE to a Seurat version 3 object}
\usage{
SCEtoSeurat(SCE, Assay = "RNA")
}
\arguments{
\item{Assay}{which assay to extract (default is "RNA")}

\item{SeuratObj}{Seurat object}
}
\description{
Function to convert Seurat object to SCE
}
\examples{
sce = SeuratToSCE(seuratObj)
}
