% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\docType{class}
\name{Seurat-class-v2}
\alias{Seurat-class-v2}
\alias{seurat}
\title{The Seurat Version 2 Class}
\description{
The Seurat object is a representation of single-cell expression data for R; each Seurat
object revolves around a set of cells and consists of one or more objects, or individual 
representations of expression data (eg. RNA-seq, ATAC-seq, etc).
These assays can be reduced from their high-dimensional state to a lower-dimension state
and stored as \code{\link{DimReduc-class}} objects. Seurat objects also store additional
meta data, both at the cell and feature level (contained within individual assays). The
object was designed to be as self-contained as possible, and easily extendible to new methods.
}
\section{Slots}{

\describe{
\item{\code{raw.data}}{A list of assays for this project}

\item{\code{meta.data}}{Contains meta-information about each cell, starting with number of genes detected (nGene)
and the original identity class (orig.ident)}

\item{\code{ident}}{The active cluster identity for this Seurat object}

\item{\code{dr}}{A list of dimmensional reduction objects for this object}

\item{\code{assay}}{A list of assays}

\item{\code{project.name}}{Name of the project}

\item{\code{misc}}{A list of miscellaneous information}

\item{\code{version}}{Version of Seurat this object was built under}

\item{\code{commands}}{A list of logged commands run on this \code{Seurat} object}

\item{\code{tools}}{A list of miscellaneous data generated by other tools, should be filled by developers only using}
}}

