% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratToCDs.R
\name{SeuratToCDs}
\alias{SeuratToCDs}
\title{Convert Seurat object to CDs for Monocle 3}
\usage{
SeuratToCDs(sObj)
}
\arguments{
\item{sObj}{}
}
\value{
cds object
}
\description{
Convert Seurat object to CDs for Monocle 3
}
\examples{
cds <- SeuratToCDs(sobj)
}
