% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratToSCE.R
\name{SeuratToSCE}
\alias{SeuratToSCE}
\title{Convert Seurat to a SCE object}
\usage{
SeuratToSCE(SeuratObj, Assay = "RNA")
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{Assay}{which assay to extract (default is "RNA")}
}
\description{
Function to convert Seurat version 3 object to SCE
}
\examples{
sce = SeuratToSCE(seuratObj)
}
