% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinPlot.R
\name{VlnPlotMarkers}
\alias{VlnPlotMarkers}
\title{Violin Plot for Seurat or SCE object}
\usage{
VlnPlotMarkers(Object, Features, FeatName = NULL, TotalExp = F,
  Clusters = NULL, Plot.Colours = NULL, Metadata = F,
  Exprs_Type = "scale.data", Assay = "RNA")
}
\arguments{
\item{Object}{either seurat object or a single cell experiment object}

\item{Features}{Gene or Gene list, or Column from Metadata/colData to plot}

\item{FeatName}{Name of the Features list. Required when gene list is is used. Will be used in the main title.Not needed for single gene.}

\item{Clusters}{Vector indicating cluster identity equal to number of cells in Object}

\item{Plot.Colours}{vector of colours with length of number of clusters.}

\item{Metadata}{T/F either gene expression (default, F) or metadata}

\item{Assay}{which seurat assay to use. Default is RNA}
}
\description{
My function for violin plots using different clusters with single genes or gene lists
}
\examples{
clusters = seuratObj@active.ident
myDimPlot(seuratObj, clusters, Reduction.Type = "tsne")
}
