% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R, R/SeuratToCDs.R
\name{SeuratToCDs}
\alias{SeuratToCDs}
\title{Convert Seurat version 3 Object to CDs for monocle}
\usage{
SeuratToCDs(sObj)

SeuratToCDs(sObj)
}
\arguments{
\item{sObj}{}

\item{SeuratObj}{Seurat object}

\item{Assay}{which assay to extract (default is "RNA")}
}
\value{
cds object
}
\description{
Function to convert Seurat object to SCE
}
\examples{
cds = SeuratToCDs(seuratObj)
cds <- SeuratToCDs(sobj)
}
