% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert.R
\name{SeuratToCDs}
\alias{SeuratToCDs}
\title{Convert Seurat version 3 Object to CDs for monocle}
\usage{
SeuratToCDs(SeuratObj, Assay = "RNA")
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{Assay}{which assay to extract (default is "RNA")}
}
\description{
Function to convert Seurat object to SCE
}
\examples{
cds = SeuratToCDs(seuratObj)
}
